Make trompeloeil::sequence moveable (#27)
* Make sequence moveable

Following up on https://github.com/rollbear/trompeloeil/pull/26#issuecomment-275454302 on the develop branch.

* Testing moveability of sequence

https://github.com/rollbear/trompeloeil/pull/26#issuecomment-275454302

* Add move assignment to sequence and list<>

* Update tests for moveability.

* Add moveability to sequence and its dependencies

* Noted the change made.

+++ trompeloeil.hpp
@@ -818,26 +818,34 @@ namespace trompeloeil
   class list_elem
   {
   public:
-    list_elem(list_elem const&) = delete;
-    list_elem& operator=(list_elem const&) = delete;
     list_elem(
       list_elem &&r)
     noexcept
-      : next(r.next)
-      , prev(&r)
     {
-      r.invariant_check();
+      *this = std::move(r);
+    }
+    list_elem& operator=(
+      list_elem &&r)
+    noexcept
+    {
+      if (this != &r)
+      {
+        next = r.next;
+        prev = &r;
+        r.invariant_check();
 
-      next->prev = this;
-      r.next = this;
+        next->prev = this;
+        r.next = this;
 
-      TROMPELOEIL_ASSERT(next->prev == this);
-      TROMPELOEIL_ASSERT(prev->next == this);
+        TROMPELOEIL_ASSERT(next->prev == this);
+        TROMPELOEIL_ASSERT(prev->next == this);
 
-      r.unlink();
+        r.unlink();
 
-      TROMPELOEIL_ASSERT(!r.is_linked());
-      invariant_check();
+        TROMPELOEIL_ASSERT(!r.is_linked());
+        invariant_check();
+      }
+      return *this;
     }
     virtual
     ~list_elem()
@@ -930,6 +938,9 @@ namespace trompeloeil
   class list : private list_elem<T>, private Disposer
   {
   public:
+    list() noexcept;
+    list(list&&) noexcept;
+    list& operator=(list&&) noexcept;
     ~list();
     class iterator;
     iterator begin() const noexcept;
@@ -1013,6 +1024,15 @@ namespace trompeloeil
   };
 
   template <typename T, typename Disposer>
+  list<T, Disposer>::list() noexcept = default;
+
+  template <typename T, typename Disposer>
+  list<T, Disposer>::list(list&&) noexcept = default;
+
+  template <typename T, typename Disposer>
+  list<T, Disposer>& list<T, Disposer>::operator=(list&&) noexcept = default;
+
+  template <typename T, typename Disposer>
   list<T, Disposer>::~list()
   {
     auto i = this->begin();
@@ -1081,7 +1101,8 @@ namespace trompeloeil
   {
   public:
     sequence() noexcept = default;
-    sequence(sequence&&) = delete;
+    sequence(sequence&&) noexcept = default;
+    sequence& operator=(sequence&&) noexcept = default;
     ~sequence();
 
     bool
